+++ nengo_dl/__init__.py
@@ -19,6 +19,11 @@ tensorflow_patch.patch_dynamic_stitch_grad()
 tensorflow_patch.patch_state_grads()
 
 # filter out "INFO" level log messages
-import tensorflow as tf  # pylint: disable=wrong-import-order,wrong-import-position
+# pylint: disable=wrong-import-order,wrong-import-position
+import os
+import tensorflow as tf
+os.environ["TF_CPP_MIN_LOG_LEVEL"] = "1"
 tf.logging.set_verbosity(tf.logging.WARN)
-del tf  # we don't want a nengo_dl.tf attribute
+# we don't want a nengo_dl.tf/os attribute
+del os
+del tf